handle check_activity for upvote and want answers
Create two new helper functions, username_correspond and want_answers.

username_correspond extract the user's username from Quora's rss endpoint (http://www.quora.com/<username/rss) and the author of the answer upvoted / written. If the username correspond it means that the entry is an answer written by the user else it is an upvoted answer.

want_answers check for the pattern # Answer(s) in the item's description.